requirements.txt
streamlit
numpy
opencv-python-headless
mediapipe
pillow
